update fire
I tweaked it wuite a bit, slowed it down a lot, made it check for water to try to spread, and I think for some reason the old version didn't care about water, this ones does.
HOWEVER: the 8 minute fire buster is no more, so we'll have to see if i need to bring it back.